requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
python-multipart>=0.0.6
sqlalchemy>=2.0.20
pydantic>=2.3.0
pillow>=10.0.0
requests>=2.31.0
